+++ src/PhpSpreadsheet/Reader/BaseReader.php
@@ -51,10 +51,11 @@ abstract class BaseReader implements IReader
      */
     protected $readFilter;
 
+    /** @var resource */
     protected $fileHandle;
 
     /**
-     * @var XmlScanner
+     * @var ?XmlScanner
      */
     protected $securityScanner;
 
@@ -134,8 +135,17 @@ abstract class BaseReader implements IReader
         return $this;
     }
 
-    public function getSecurityScanner()
+    public function getSecurityScanner(): ?XmlScanner
     {
+        return $this->securityScanner;
+    }
+
+    public function getSecurityScannerOrThrow(): XmlScanner
+    {
+        if ($this->securityScanner === null) {
+            throw new ReaderException('Security scanner is unexpectedly null');
+        }
+
         return $this->securityScanner;
     }