+++ src/PhpSpreadsheet/Reader/Ods/Properties.php
@@ -8,6 +8,7 @@ use SimpleXMLElement;
 
 class Properties
 {
+    /** @var Spreadsheet */
     private $spreadsheet;
 
     public function __construct(Spreadsheet $spreadsheet)
@@ -15,7 +16,7 @@ class Properties
         $this->spreadsheet = $spreadsheet;
     }
 
-    public function load(SimpleXMLElement $xml, $namespacesMeta): void
+    public function load(SimpleXMLElement $xml, array $namespacesMeta): void
     {
         $docProps = $this->spreadsheet->getProperties();
         $officeProperty = $xml->children($namespacesMeta['office']);
@@ -69,9 +70,9 @@ class Properties
     }
 
     private function setMetaProperties(
-        $namespacesMeta,
+        array $namespacesMeta,
         SimpleXMLElement $propertyValue,
-        $propertyName,
+        string $propertyName,
         DocumentProperties $docProps
     ): void {
         $propertyValueAttributes = $propertyValue->attributes($namespacesMeta['meta']);
@@ -96,6 +97,10 @@ class Properties
         }
     }
 
+    /**
+     * @param mixed $propertyValueAttributes
+     * @param mixed $propertyValue
+     */
     private function setUserDefinedProperty($propertyValueAttributes, $propertyValue, DocumentProperties $docProps): void
     {
         $propertyValueName = '';