+++ tests/PhpSpreadsheetTests/Reader/Gnumeric/AutoFilterTest.php
@@ -3,29 +3,20 @@
 namespace PhpOffice\PhpSpreadsheetTests\Reader\Gnumeric;
 
 use PhpOffice\PhpSpreadsheet\Reader\Gnumeric;
-use PhpOffice\PhpSpreadsheet\Spreadsheet;
 use PHPUnit\Framework\TestCase;
 
 class AutoFilterTest extends TestCase
 {
-    /**
-     * @var Spreadsheet
-     */
-    private $spreadsheet;
-
-    protected function setUp(): void
+    public function testAutoFilterRange(): void
     {
         $filename = 'tests/data/Reader/Gnumeric/Autofilter_Basic.gnumeric';
         $reader = new Gnumeric();
-        $this->spreadsheet = $reader->load($filename);
-    }
-
-    public function testAutoFilterRange(): void
-    {
-        $worksheet = $this->spreadsheet->getActiveSheet();
+        $spreadsheet = $reader->load($filename);
+        $worksheet = $spreadsheet->getActiveSheet();
 
         $autoFilterRange = $worksheet->getAutoFilter()->getRange();
 
         self::assertSame('A1:D57', $autoFilterRange);
+        $spreadsheet->disconnectWorksheets();
     }
 }